requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
anthropic==0.8.1
python-dotenv==1.0.0
flask==2.3.3